+++ iti-react/DataUpdater/DataUpdater.tsx
@@ -1,11 +1,6 @@
 ﻿import { debounce } from 'lodash'
 import { CancellablePromise } from '../CancellablePromise'
 
-// String of random characters at the end to avoid unintentional collisions,
-// if TQueryParams is string.
-const notProvided = 'not-provided-_JCDD*{J`mmQ_R4z'
-type NotProvided = 'not-provided-_JCDD*{J`mmQ_R4z'
-
 export interface IDataUpdater<TQueryParams> {
     doQuery(changeLoading?: boolean): void
     doQueryAsync(changeLoading?: boolean): Promise<void>
@@ -61,26 +56,26 @@ export class DataUpdater<TQueryParams, TResult> implements IDataUpdater<TQueryPa
         this.onError = options.onError
 
         this.doQueryDebounced = debounce(
-            (changeLoading?: boolean) => this.doQueryInternal(notProvided, changeLoading),
+            (changeLoading?: boolean) =>
+                this.doQueryInternal(undefined, false, changeLoading),
             500
         )
     }
 
-    // optionalQueryParams allows the string 'not-provided', because queryParams can
-    // legitimately be undefined or null. So, we need a different way to check if the
-    // optionalQueryParams argument was supplied or not.
-    //
-    //
+    // optionalQueryParamsProvided exists because the queryParams can legitimately be
+    // null or undefined, so we need a different way to tell if the caller of the method
+    // provided a value for optionalQueryParams.
     private async doQueryInternal(
-        optionalQueryParams: TQueryParams | NotProvided,
+        optionalQueryParams: TQueryParams | undefined,
+        optionalQueryParamsProvided: boolean,
         changeLoading: boolean = true,
         handleErrors: boolean = true
     ) {
         this.onQueryStarted()
 
         let queryParams: TQueryParams
-        if (optionalQueryParams !== notProvided) {
-            queryParams = optionalQueryParams
+        if (optionalQueryParamsProvided) {
+            queryParams = optionalQueryParams!
         } else {
             queryParams = this.getCurrentQueryParams()
         }
@@ -117,11 +112,11 @@ export class DataUpdater<TQueryParams, TResult> implements IDataUpdater<TQueryPa
     }
 
     doQueryAsync = async (changeLoading: boolean = true) => {
-        await this.doQueryInternal(notProvided, changeLoading, false)
+        await this.doQueryInternal(undefined, false, changeLoading, false)
     }
 
     doQuery = (changeLoading: boolean = true) => {
-        this.doQueryInternal(notProvided, changeLoading)
+        this.doQueryInternal(undefined, false, changeLoading)
     }
 
     handleQueryParamsChange(
@@ -132,7 +127,7 @@ export class DataUpdater<TQueryParams, TResult> implements IDataUpdater<TQueryPa
         if (shouldDebounce) {
             this.doQueryDebounced(changeLoading)
         } else {
-            this.doQueryInternal(queryParams, changeLoading)
+            this.doQueryInternal(queryParams, true, changeLoading)
         }
     }